pyproject.toml
[project]
name = "oww-trainer"
version = "0.1.0"
description = "openWakeWord model trainer - automated wakeword model training pipeline"
requires-python = ">=3.10"
dependencies = [
    "openwakeword",
    "torch>=2.0,<2.5",
    "torchaudio>=2.0,<2.5",
    "torchinfo>=1.8.0,<2",
    "torchmetrics>=1.0,<2",
    "numpy<2",
    "scipy>=1.3,<1.15",
    "pyyaml>=6.0,<7",
    "datasets>=2.14,<3",
    "mutagen>=1.46,<2",
    "speechbrain>=0.5.14,<1",
    "audiomentations>=0.30,<1",
    "torch-audiomentations>=0.11,<1",
    "acoustics>=0.2.6,<1",
    "pronouncing>=0.2.0,<1",
    "deep-phonemizer==0.0.19",
    "piper-phonemize",
    "webrtcvad",
    "tqdm>=4.64,<5",
    "rich>=13,<14",
    "setuptools>=70,<76",
    "espeak-phonemizer>=1.3.1",
    "onnx>=1.14,<1.18",
    "silero-vad>=5.1,<6",
    "sounddevice>=0.4,<1",
    "loguru>=0.7,<1",
    "piper-tts>=1.4.1",
]

[dependency-groups]
dev = [
    "ruff",
    "ty",
    "pre-commit",
    "pytest",
    "pytest-mock",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/oww_trainer"]

[tool.uv.sources]
openwakeword = { path = "../openWakeWord", editable = true }

[tool.ruff]
line-length = 120
target-version = "py310"
src = ["src"]
exclude = ["src/oww_trainer/train.py"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
